Fixes for generated datapoints table:
- Structured input rendering to match data table, (moving the formatForDisplay() method to types.ts so it can be used in both places)

- "table widths are weird" bug from the data table (currently, the generated datapoints table does not use table.ts, so this moves that logic into types.ts as well so it can be used in both places.)

- Combining "counterfactuals" and "counterfactual explanations" tabs (+ tweaks to both)

PiperOrigin-RevId: 342308747
