+++ lit_nlp/client/lib/types_test.ts
@@ -17,41 +17,40 @@
 
 import 'jasmine';
 
-import {DataTableModule} from './data_table_module';
+import {formatForDisplay} from './types';
 
 describe('formatForDisplay test', () => {
-  const dataTableModule = new DataTableModule();
 
   it('formats string for display', () => {
     const testString = 'test';
-    const formatted = dataTableModule.formatForDisplay(testString);
+    const formatted = formatForDisplay(testString);
     expect(formatted).toBe(testString);
   });
 
   it('formats number for display', () => {
-    const formatted = dataTableModule.formatForDisplay(1.23456789);
+    const formatted = formatForDisplay(1.23456789);
     expect(formatted).toBe('1.2346');
   });
 
   it('formats boolean for display', () => {
-    let formatted = dataTableModule.formatForDisplay(true);
+    let formatted = formatForDisplay(true);
     expect(formatted).toBe('✔');
 
-    formatted = dataTableModule.formatForDisplay(false);
+    formatted = formatForDisplay(false);
     expect(formatted).toBe(' ');
   });
 
   it('formats array for display', () => {
     // number array.
-    let formatted = dataTableModule.formatForDisplay([1.23456789, 2.3456789]);
+    let formatted = formatForDisplay([1.23456789, 2.3456789]);
     expect(formatted).toBe('1.2346, 2.3457');
 
     // number|string array.
-    formatted = dataTableModule.formatForDisplay(['a', 1.23456789, 2.3456789]);
+    formatted = formatForDisplay(['a', 1.23456789, 2.3456789]);
     expect(formatted).toBe('a, 1.2346, 2.3457');
 
     // string array
-    formatted = dataTableModule.formatForDisplay(['a', 'b', 'cd']);
+    formatted = formatForDisplay(['a', 'b', 'cd']);
     expect(formatted).toBe('a, b, cd');
   });
 });
