+++ lit_nlp/client/modules/salience_map_module.css
@@ -55,3 +55,8 @@
   align-items: center;
   justify-content: center;
 }
+
+.config {
+  display: flex;
+  align-items: center;
+}
